Fix for #18 - improve cartography activity assignment and also added tests for recentjournal.py plus refatcoring and code tidy up
